batch-etl-validate-postgres/requirements.txt
psycopg[binary]>=3.1.18
pandas>=2.2.0
pandera>=0.20.0
typer>=0.12.0
rich>=13.0.0
python-dotenv>=1.0.0

pytest>=8.0.0
ruff>=0.5.0
